Gemfile.lock: bundle update
There will be no changes, since cinch 2.3.2 doesn't change anything that
we test, nor does cinch-test 0.1.3, which only affects tests with
private messages to the bot. We do have tests sending private messages,
but we never tested the specific number of the replies, only that they
are non-empty and/or match other private message responses.